+++ src/avista_digital_exchange_sdk/dataCaptureUtil.py
@@ -1,6 +1,6 @@
 
 from logging import Logger
-from typing import List
+from typing import AsyncIterator, List
 from .async_graphql_client import exceptions, input_types
 from .client import Client as DxClient
 from . import DxTypes
@@ -16,13 +16,16 @@ class DataCaptureUtil(object):
         self._client = client
         self._logger = logger
 
-    async def listenForCaptureData(self, captureId: str):
-        print("Subscribing to capture data....")
+    async def listenForCaptureData(self, captureId: str) -> AsyncIterator[DxTypes.PublishCaptureDataResult]:
+        if self._debug:
+            print("Subscribing to capture data....")
         async for item in self._client.updatedGqlClient.on_capture_publish_data(
                 capture_id=captureId):
-            print("In for loop...")
-            print("Subscription item....")
-            print(item)
+            if self._debug:
+                print("DataCapture.listenForCaptureData: Data received.")
+            result = DxTypes.PublishCaptureDataResult.fromCapturePublishResult(
+                item.on_capture_publish_data)
+            yield result
 
     async def publishData(self, captureId: str, data: List[DxTypes.CaptureDataRecordInput]) -> DxTypes.PublishCaptureDataResult:
         if self._debug: